Reorganized folder structure to follow 'openshift-quickstar' pipeline; package.json updated
